glycan-binding-site-predictor/requirements.txt
torch
fair-esm
